requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.1
Django-ddd==0.1.10
django-rest-swagger==2.2.0
djangorestframework==3.14.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.5
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
openapi-codec==1.3.2
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
